SentimentAnalysis/requirements.txt
django
mysqlclient
python-dotenv
django-safedelete
mysql-client
djangorestframework
transformers
torch